requirements.txt
flask>=3.0
pandas>=2.1
openpyxl>=3.1
Pillow>=10.0
PyYAML>=6.0
numpy>=1.24
torch>=2.1
torchvision>=0.16
ftfy>=6.1
regex>=2023.0
tqdm>=4.66
open-clip-torch>=2.24
